ChopBlanks should not trim binary fields
Binary fields are incorrectly being trimmed of trailing spaces when non
server side prepare is being used. Trimming of server side prepare binary
fields is catered for properly.

"As of MySQL 5.0.3, trailing spaces are retained when values are stored and
retrieved, in conformance with standard SQL. Before MySQL 5.0.3, trailing
spaces are removed from values when they are stored into a VARCHAR column;
this means that the spaces also are absent from retrieved values."

This means that testing VARCHAR (and VARBINARY as it turns out) on 4.1.22
will cause the test cases to fail. However, since we're testing BLOB,
TINYBLOB, MEDIUMBLOB and LONGBLOB, we have enough binary fields to confirm
that after the change we do not trim the trailing spaces from binary fields
after retrieval on the client side.

Fixes: https://github.com/perl5-dbi/DBD-mysql/pull/138