Properly set STMT_ATTR_UPDATE_MAX_LENGTH attribute when needed
STMT_ATTR_UPDATE_MAX_LENGTH attribute is needed to set for statements in
case at least one column contains non-primitive type (that which has no
fixed length). When this attribute is not set then database server does not
send maximal length of variable column. And so driver would not have enough
large buffer for storing column result.

Unify also logic for determining if result would be stored in perl's PV
slot or IV/NV slot.

This patch fixes also zerofill support for MariaDB 10.2.6+ which did not
work because of unknown length of buffer.

Fixes: https://github.com/perl5-dbi/DBD-mysql/pull/133