Ensure that MYSQL_VERSION_ID and MARIADB_BASE_VERSION are defined
Macro MYSQL_VERSION_ID needs to be always set as driver has a lot of ifdefs
with this macro. And it is needed also when compiling with MariaDB. In some
client versions it is not set by main mysql.h header file and for proper
definition is needed to include mysql_version.h header file.

On the other hand macro MARIADB_BASE_VERSION is used for determining if
compiling with MariaDB or not. But in some cases MariaDB does not set this
macro which leads to incorrect compilation and broken DBD::MariaDB driver.
So define MARIADB_BASE_VERSION based on another MARIADB_PACKAGE_VERSION
which seems to be present when MARIADB_BASE_VERSION is not.

This should fix compilation with MariaDB Connector/C of version 3.

Fixes: https://github.com/perl5-dbi/DBD-mysql/issues/139
Fixes: https://github.com/perl5-dbi/DBD-mysql/pull/133