Hide Makefile.PL warnings from `` operator
When mariadb_config is not available in the system, then perl throw
following warning:

Can't exec "mariadb_config": No such file or directory at Makefile.PL line 81.

And this warning is thrown even if mysql_config is available. Because this
code is there just for checking and determining which application is
installed and available in the system it does not make sense to throw any
warning.

Use 2>&1 expression to hide that warning. `` still returns undef when
command is not possible to execute. Expression 2>/dev/null cannot be used
as it is not supported on Windows systems.