Remove code which parses LIMIT and CALL substrings in SQL statement
This code tried to switch from server side prepare statement to client side
statements with emulation for placeholders after LIMIT or CALL keywords.

But it was incorrect as it used check version of client library instead of
MySQL server version. So it did not work. Moreover there is already code
for fallback, therefore this check for LIMIT or CALL substring is not
needed at all.

Fallback code is extended for error ER_PARSE_ERROR which is returned in
this case by older MySQL servers.