Fix checking if bind value is mysql number and if needs to be quoted
Before this patch was used function parse_number() which said that strings
'.', '+' or '-' are valid numbers and so does not have to be quoted.

As driver code just needs to know if value is number or not, parse_number()
function was renamed to is_mysql_number() and rewritten to accept only
string which is valid mysql number.

Because number of placeholders (?) must match number of bind parameters
(this is already checked) it is not needed to check/detect type of bind
parameter two times. So first detection was removed.