In Makefile.PL properly specify libdirs for Devel::CheckLib
Devel::CheckLib process libdirs correctly only when they are specified in
'LIBS' or 'libpath' argument. This fixes checking for client libraries if
they are located in non-standard or non-system location.

Due to way how Devel::CheckLib process header files, it is needed to have
libdirs specified also in ldflags in case there are some helper libraries
specified in ldflags.