Allow to specify custom --language parameter for test t/12embedded.t
This parameter is needed for running tests on Travis as language directory
on Travis is on non-standard location. Specify this parameter via env
variable DBD_MARIADB_TESTLANGDIR and set it in Travis configuration.

Also it fixes problems when it was not possible to start embedded server.
So remove skip for this error message. For this test it is fatal error.