Detect mysql_get_option() function at configure time via check_lib()
It looks like that this function is missing in some MariaDB configurations.
So rather then trying to guess MariaDB versions in which it is missing,
detect presence of mysql_get_option() function at configure time in
Makefile.PL via Devel::CheckLib's check_lib().

Also when function mysql_get_option() is not available then it is not
possible to fetch value of $dbh->{mariadb_max_allowed_packet}. So properly
throw DBI error.