Process autoreconnect in mariadb_st_internal_execute41() function
If autoreconnect is enabled then it initialize a new connection to database
server. If autoreconnect fails or when is disabled then this function
immediately returns and throw an error.

This ensures that connection to database server is active prior to
executing prepared statement and also prevents possible NULL pointer
dereference as mysql_* functions needs to have valid MYSQL* connection
structure.