Check that correct number of bind parameters were supplied for $dbh->do() method
Same change as in commit 989bdb3cb27ff67806f50ebf76f390ae465fd4dd but this
time for DBI's $dbh->do() method.

When number of bind parameters was less then number of placeholders in
statement then DBD::MariaDB caused SEGFAULT of whole perl process as it
read out-of-bound memory. Check in this change prevents crashing of perl
process and properly propagates DBI error to the caller.