Cover a protocol design error for multi result statements
Error packet does not contain the server status, so client does not know if
there is another multi result or not. Luckily, an error always aborts
execution of a statement, so it is safe to expect that there is no result
statement and turn off the SERVER_MORE_RESULTS_EXISTS flag.

Applying this workaround for protocol error is needed after every failed
call which may produce multi result set and we already received some multi
result set. So after mysql_next_result() or after mysql_real_query() with
multi statements.