Do not change prepared statement row_num when there are no more data
Function mysql_stmt_affected_rows() returns number of affected rows for the
UPDATE, DELETE or INSERT statement. So it does not make sense to call this
function for statement with result set.

Also function mariadb_st_execute_iv() already set correct number for
row_num (either number of affected rows or number of rows in result set) so
there is absolutely no need to update it in mariadb_st_fetch() function.

Seems that this code has not caused any problems as documentation for
mysql_stmt_affected_rows() function says:

  For a description of what the affected-rows value returned by this
  function means, see mysql_affected_rows().

And for mysql_affected_rows():

  For SELECT statements, mysql_affected_rows() works like mysql_num_rows().

And for mysql_num_rows():

  Returns the number of rows in the result set.