Run AppVeyor CI on the correct pull request commit
AppVeyor is testing merge commit of the pull request and not the commit for
which was CI triggered. The correct commit for which was CI triggered is
stored in env variable %APPVEYOR_PULL_REQUEST_HEAD_COMMIT%. Variable is
empty for regular non-pull requests CI tests. Calling git checkout without
commit/branch does nothing.