Fix $sth->finish() for mariadb_use_result
mysql_use_result() requires fetching all rows before executing the
next query, otherwise we'd get an error 2014 (Commands out of sync).

This can be relevant if the script leaves the result loop early. If it
explicitly calls $sth->finish() that would croak, if it undefines $sth
it would get a warning. If it doesn't do anything and executes another
query, it croaks. With this fix, the pending rows are fetched silently.

For multi-result sets we could use mysql_store_result() for the
remaining results, that would consume more memory to buffer all rows,
rather than just fetching and discarding each individual row.