Skip testing the old server with a new client to prevent crashes
It happens when the client is parsing response for the PREPARE STATEMENT
packet which contains PARAMETERS metadata.

After all PARAMETERS metadata is the EOF packet. But the MySQL 4.1
server does not return in this EOF packet Warning Count and Server
Status fields.

MySQL 8.0+ client code in the function cli_read_metadata_ex() which is
called from cli_read_prepare_result(), does not check for the packet
length and hence try to dereference memory out of the buffer and assigns
into the mysql->server_status variable some garbage data.