Skip double socket close prevention for Windows UCRT builds
New Windows UCRT library which implements C Runtime and is replacement for
old MSVCRT library does not export __pioinfo[] array and so does not
provide a way to replace native Windows handle associated with file
descriptor. As the __pioinfo[] array symbol is not available, compilation
of DBD-MariaDB is causing linker errors when building with Perl compiled
against UCRT (as opposite of the MSVCRT).

Skip the double socket close prevention for Windows UCRT builds as it is
not possible to compile this code.

Fixes: https://github.com/perl5-dbi/DBD-MariaDB/issues/231